factor_mayor decimal: soportar empaques fraccionados (Doc v4)
Ajuste v4.0 (Compras al Mayor por Bulto, Caja y Rema): el factor de conversión
debe admitir decimales por si existen empaques fraccionados.

- conceptos.factor_mayor pasa de INT a DECIMAL(10,2) DEFAULT 1.00. El auto-
  instalador migra automáticamente la columna existente (INT -> DECIMAL) además
  de crearla ya decimal en instalaciones nuevas (nuevo paso 3d + helper
  columnaDataType()).
- insertar/actualizar_concepto.php: el factor se toma con floatval + round(.,2)
  (antes intval); sigue cayendo a 1 si es inválido o <= 0.
- guardar_egreso.php: unidades_netas = round(cantidad_presentaciones × factor)
  al reponer inventario (el stock se lleva en enteros). Ej: 3 cajas × 12.5 = 38 u.
- Productos: la entrada del factor admite decimales (step 0.01) y el texto de
  ayuda menciona bulto/caja/rema/paquete. Render de tabla y tarjeta del POS de
  egresos usan parseFloat (12.50 se muestra 12.5; 500.00 se muestra 500).

Verificado end-to-end: migración INT->DECIMAL(10,2); factor 12.5 guardado;
3 cajas × 12.5 => stock +38; factor entero 500 intacto (500.00).

+++ Felix_sistema/assets/js/facturacion.js
@@ -80,7 +80,7 @@ function initPOS(config) {
             // En el POS de egresos (compras al mayor) se avisa cuántas unidades
             // al detal repone cada presentación, para que quede claro que la
             // cantidad se cuenta en presentaciones (ej. Resmas), no en hojas.
-            const factor = parseInt(p.factor_mayor) || 1;
+            const factor = parseFloat(p.factor_mayor) || 1;
             const factorTxt = (config.esMayorista && esProducto && factor > 1)
                 ? `<span class="tile-stock">1 pres = ${factor} u</span>` : '';
             const btn = document.createElement('button');
